[BUGFIX] Fixed Write Callback
This is a small bug with huge significance right under my nose. The
write callback was not resizing raw data buffer in parsed response
object properly.

Due to this bug, the program would make attempts to parse incomplete
response data on slow internet because data received is in multiple chunks.

This commit fixes that.

TODO: `DELETE : /v1/analyse/{binary_id}` endpoint is still not working properly.