[HACK][BUG] Fix Delete & Analysis Status Endpoints
Bug Description:
For some time, a bug caused calls to specific API endpoints
to fail. Examples are `DELETE {host}/analysis/{binary_id}` and `GET
{host}/analysis/status/{binary_id}`. This was probably due to a
`HTTP/1.1 100-continue` received from the server and the curl waiting
for a large chunk of data to be retrieved, but after that it hangs.

The behaviour can be observed when `CURLOPT_VERBOSE` is enabled.

Hack:
This commit fixes the bug by resetting the curl handle. Since I have to
move fast, and spent significant amount of my time to fix this (and was
not able to), I'm using this hack for now and leave the debugging to later.