Fix External C Linkage, Fix tomlc99 Installation
- Previously `extern "C"` wasn't working properly because of some syntax
errors, this commit fixes that and removes `C_SOURCE_BEGIN` and
`C_SOURCE_END` and explicitly replaces that with proper extern "C".

- Whe library was installed and used in rizin plugin, due to linkage
failures, rizin was unable to load plugin. This commit fixes that by
properly adding linkage with libtomlc99 and properly installing that
library as well when `creait` is installed.