Single & Batch Rename API
This commit adds single and batch rename API & also adds convinient
wrappers for them.

A bug in parsing of response for `GET basic_function_info` was also
fixed. The bug involved the length of fn_info vec not being updated as
new items are inserted. This commit fixes that. This bug was not
encountered before.