ReaiDb Integration With Reai Object
Reai can now directly use ReaiDb in order to update the database with
relevant information automatically.

One more API is added to Reai object. The call can be used to update
analysis status of all previously created analysis in database. This is
useful because initially the analysis status is set to Queued.