ReaiLog Works
Turns out that ReaiLog worked from the starting, but in the process
maybe I found an interesting bug that I'll spend my time exploring this
weekend.

There was no problem in the Log from the start and the problem was only
in the usage. Now that does not exist, so we're clear.