Test Function Similarity & Fix Database Query
Previous commit was untested for the new API endpoint added,
now it's used in an example and it seems to be working just fine.

Slight modification to the database made in order to sort the results
properly. When getting analysis, whether through get_all or get_latest,
the results are sorted in descending order by creation_date_time column.

This means all methods will return results in a sorted order.