Remove Local Database Feature
Local database management turned out to not be such a good idea. Why?
because now users can apply existing analysis in rizin and cutter
plugins and hence no need to track binary id or analysis id or hashes of
uploaded files.

If this new feature were not to be added, then local db management is a
very helpful feature to have IMO