[HACK] Add CI Tests & Make Up For Poor Documentation
Documentation for `/v1/ann/symbol/batch` seems to be incorrect.
Can't wait for getting a response and then making decision based
on that, so for now moving ahead with a hack to optionally check
for `function_matches` field the second time. I'm referring
to the JSON response parsing code in Response.c