[SQUASH] Add Required `v1` API Features With Tests & CI (#2)
* marking v1

* Debug Mode

* Remove Local Database Feature

Local database management turned out to not be such a good idea. Why?
because now users can apply existing analysis in rizin and cutter
plugins and hence no need to track binary id or analysis id or hashes of
uploaded files.

If this new feature were not to be added, then local db management is a
very helpful feature to have IMO

* Log Now Available Globally

- Auth Check endpoint works
- Log object management removed, Log now available globally. Logging
will be initialized as soon as the library is loaded and all logs will
be stored in tmpdir.

* Auth Check Takes Host Param As Well

API key works correctly only with it's corresponding host. Auth check
now sets the host as well.

* Get AI Models

Add new endpoint to get available AI models

* Remove Redundant Model Arguments

Plugins now don't use a static modelname-version syntax to specify
versions. All AI model information is passed into the struct of
arguments of API that requires it.

* Mock API

* [BUG] Add Tests

There are some bugs found during execution of these tests
Strange, I never thought they'd exist like this. While writing
the tests I thought they'd be useless.

* Fix Some Memory Leaks

The bug still remains unfixed.

* [BUGFIX] Remove Free On Json String

The json string getter method does not duplicate the string
and hence ownership still remained with cJSON object. One of
the Json getter method destroyed the returned string and hence
it was a double-free bug, because in the end the cJSON object
tried to destroy it was well.

Phew!

* Update README For cJSON Dependency

Recently faced a RPATH issue when building and installing cJSON from
source on Mac OS. The updated README provides hints on how that can be
fixed.

* [HACK] Add CI Tests & Make Up For Poor Documentation

Documentation for `/v1/ann/symbol/batch` seems to be incorrect.
Can't wait for getting a response and then making decision based
on that, so for now moving ahead with a hack to optionally check
for `function_matches` field the second time. I'm referring
to the JSON response parsing code in Response.c

* Update ci.yaml To Run On Any Commit/Push

* CI Error Fix

* Update CI To Create TMP Directory & Display Logs

* To Review & Merge

Some small non-necessary fixes to follow a proper
coding style throughout.

* Use `memmove` in vec_remove

* `FREE` Automatically Sets To `NULL`

No need to explicitly sets variables to `NULL` after using `FREE` on
them, this is automatically done in `FREE` macro