New API Endpoint To Get Basic Collection Info
API Endpoint to get similar functions takes collection ids. From the
user's perspective getting a collection id from a search either through
the Web UI or through the plugin (when it supports that feature) can be
a bad UX.
For this, whenever an array of collection IDs is required, this endpoint
can be used to search for a general set of collection by name, and then
taking the top 10 or 20 or any N results and providing collection IDs.